Recode of the entire package selection
Instead of mucking around with lots of state variables, keep track
of state in a linked list, controlled by a StateManager. This provides
better caching, encapsulation of pkg-query(8) use and it gives us
redo. The code is naturally longer, due to additional boilerplate,
but it makes a lot more sense and was actually faster written than
the original implementation.